# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.DefaultResponse:
    properties:
      message:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controllers.PasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.SpecialistActivityReport:
    properties:
      application_count:
        type: integer
      full_name:
        type: string
    type: object
...
  models.SwagUser:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      username:
        type: string
    type: object
...
  models.User:
    properties:
      birth_date:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: This is a Tajik Career Hub API documentation.
  title: Tajik Career Hub API ✨
  version: "1.0"
paths:
  /activities:
    get:
      consumes:
      - application/json
      description: Get a report of how many vacancies a specific specialist has applied
        for
      operationId: get-specialist-activity-report-by-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SpecialistActivityReport'
            type: array
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get specialist activity report for a specific user
      tags:
      - Reports
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users or a single user by username
      parameters:
      - description: Username to filter the user
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Single user details
          schema:
            $ref: '#/definitions/models.User'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.SwagUser'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user with the provided details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SwagUser'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid ID or input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user details
      tags:
      - Users
  /users/block/{id}:
    patch:
      consumes:
      - application/json
      description: Block a user by their ID
      parameters:
      - description: User ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User blocked successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Block user
      tags:
      - Admin
  /users/password:
    patch:
      consumes:
      - application/json
      description: Update the password for the current user. Requires authentication.
      parameters:
      - description: Password request data
        in: body
        name: passwordRequest
        required: true
        schema:
          $ref: '#/definitions/controllers.PasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user password update successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user password
      tags:
      - Users
  /users/unblock/{id}:
    patch:
      consumes:
      - application/json
      description: Unlock a user by ID
      parameters:
      - description: User ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User unblocked successfully
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unblock user
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"